+++ code/game-library-proto/app.js
@@ -1,10 +1,6 @@
 const express = require('express');
 const app = express();
 
-// using bcrypt to hash and compare passwords
-// from https://medium.com/@holtkam2/add-user-authentication-to-your-node-expressjs-application-using-bcrypt-81bb0f618ab3
-const bcrypt = require('bcrypt');
-
 // check for config.js
 // from https://stackoverflow.com/questions/4482686/check-synchronously-if-file-directory-exists-in-node-js/4482701
 const fs = require('fs');
@@ -17,6 +13,24 @@ const routes = require('./routes/root');
 
 // using the ejs template engine
 app.set('view engine', 'ejs');
+
+// use express-sessions
+const session = require('express-session');
+app.use(session({
+    secret: config.session.password,
+    resave: false,
+    saveUninitialized: false,
+}));
+
+// use body-parser
+const bodyParser = require('body-parser');
+app.use(bodyParser.urlencoded({extended: true})); 
+
+// add session to context
+app.use(function(req,res,next){
+    res.locals.session = req.session;
+    next();
+});
 
 // serve static files in the ./public folder
 app.use(express.static('public'));